Simplified //hadoop/formats/util class
- Collapsed JanusGraphHadoop interface and implementation, and renamed to JanusHadoopSetup
- Moved JanusHadoopSetup construction to happen in JanusGraphVertexDeserialiser
- Renamed JnausGraphVertexDeserialiser to be VertexDeserialiser
- Moved JanusHadoopSetup class to be nested in VertexDeserialiser
- Moved SystemTypeInspector class to be nested in VertexDeserialiser
